Transcription Service
The backend class to handle Speech-to-Text requested by VoiceRecorderButton.
Takes a Base64Sound String.

Groq  Speech-to-Text does not work on AAC audio format so for android (non web) platform, this script attempts to convert the string to AAC type blob and turn that into WAV blob.

For web, the string already represents WAV format, so we directly obtain a WAV blob.

The WAV is sent to groq-sdk and text is obtained.

[Improvements required => AAC to WAV conversion algorithm
Groq api key input]